# docs/api/openapi-v1.json
{
  "components": {
    "schemas": {
      "AcuteSymptomsSchema": {
        "additionalProperties": false,
        "description": "Public mirror of the deterministic acute symptom/onset contract.",
        "properties": {
          "arm_weakness": {
            "default": false,
            "title": "Arm Weakness",
            "type": "boolean"
          },
          "balance_or_coordination_loss": {
            "default": false,
            "title": "Balance Or Coordination Loss",
            "type": "boolean"
          },
          "confusion_or_understanding_difficulty": {
            "default": false,
            "title": "Confusion Or Understanding Difficulty",
            "type": "boolean"
          },
          "face_drooping": {
            "default": false,
            "title": "Face Drooping",
            "type": "boolean"
          },
          "speech_difficulty": {
            "default": false,
            "title": "Speech Difficulty",
            "type": "boolean"
          },
          "sudden_severe_headache": {
            "default": false,
            "title": "Sudden Severe Headache",
            "type": "boolean"
          },
          "symptom_onset_minutes": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symptom Onset Minutes"
          },
          "symptoms_resolved": {
            "default": false,
            "title": "Symptoms Resolved",
            "type": "boolean"
          },
          "vision_disturbance": {
            "default": false,
            "title": "Vision Disturbance",
            "type": "boolean"
          }
        },
        "title": "AcuteSymptomsSchema",
        "type": "object"
      },
      "AssessmentCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "acute_symptoms": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AcuteSymptomsSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "audio_attachment_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Attachment Id"
          },
          "case_id": {
            "format": "uuid",
            "title": "Case Id",
            "type": "string"
          },
          "face_attachment_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Face Attachment Id"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "metadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MetadataInputSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "session_id": {
            "default": "api-session",
            "maxLength": 128,
            "minLength": 1,
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "case_id"
        ],
        "title": "AssessmentCreateRequest",
        "type": "object"
      },
...
      "AssessmentResponse": {
        "additionalProperties": false,
        "properties": {
          "case_id": {
            "format": "uuid",
            "title": "Case Id",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "request_snapshot": {
            "additionalProperties": true,
            "title": "Request Snapshot",
            "type": "object"
          },
          "result_snapshot": {
            "additionalProperties": true,
            "title": "Result Snapshot",
            "type": "object"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "case_id",
          "status",
          "request_snapshot",
          "result_snapshot",
          "created_at"
        ],
        "title": "AssessmentResponse",
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "correlation_id": {
            "title": "Correlation Id",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "details",
          "correlation_id"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "MetadataInputSchema": {
        "additionalProperties": false,
        "description": "Public mirror of the canonical ten-field metadata adapter contract.",
        "properties": {
          "Residence_type": {
            "anyOf": [
              {
                "enum": [
                  "Rural",
                  "Urban"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Residence Type"
          },
          "age": {
            "anyOf": [
              {
                "maximum": 120.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "avg_glucose_level": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Glucose Level"
          },
          "bmi": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bmi"
          },
          "ever_married": {
            "anyOf": [
              {
                "enum": [
                  "No",
                  "Yes"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ever Married"
          },
          "gender": {
            "anyOf": [
              {
                "enum": [
                  "Female",
                  "Male",
                  "Other"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "heart_disease": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heart Disease"
          },
          "hypertension": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hypertension"
          },
          "smoking_status": {
            "anyOf": [
              {
                "enum": [
                  "Unknown",
                  "formerly smoked",
                  "never smoked",
                  "smokes"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Smoking Status"
          },
          "work_type": {
            "anyOf": [
              {
                "enum": [
                  "Govt_job",
                  "Never_worked",
                  "Private",
                  "Self-employed",
                  "children"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Work Type"
          }
        },
        "title": "MetadataInputSchema",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "RuralStroke-Assist API",
    "version": "0.0.1-stage2"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/assessments": {
      "post": {
        "operationId": "assessment_create",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssessmentCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid assessment request."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Assessment unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Assessment",
        "tags": [
          "assessments"
        ]
      }
    },
    "/api/v1/assessments/{assessment_id}": {
      "get": {
        "operationId": "assessment_get",
        "parameters": [
          {
            "in": "path",
            "name": "assessment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Assessment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Dependency unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Assessment",
        "tags": [
          "assessments"
        ]
      }
    },
    "/api/v1/assessments/{assessment_id}/result": {
      "get": {
        "operationId": "assessment_result_get",
        "parameters": [
          {
            "in": "path",
            "name": "assessment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Assessment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal server error."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Dependency unavailable."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Assessment Result",
        "tags": [
          "assessments"
        ]
      }
    },
...
  }
}